Refactoring into app paterns
