+++ src/frontend/index.css
@@ -8,9 +8,3 @@ html,body,#app {
   width: 100%;
   height: 100%;
 }
-
-.hidden-ui {
-  opacity: 0;
-  pointer-events: none;
-  transition: opacity 120ms linear;
-}
